Change getter and setter for type in AvailabilityTypes Moved test for PairResource to PairResourceTests package
