+++ src/main/java/resources/PairResource.java
@@ -23,7 +23,6 @@ import java.util.stream.Stream;
 @Produces(MediaType.APPLICATION_JSON)
 public class PairResource {
 
-    private final String AVAILABLE_TYPE = "available";
     private SlotsDao slotsDao;
 
     @Inject
@@ -141,7 +140,7 @@ public class PairResource {
                         .stream()
                         .filter(ss -> rs.getSlot().equals(ss.getSlot())
                                 && rs.getSlotsDate().equals(ss.getSlotsDate())
-                                && rs.getType().getType().equals(AvailabilityTypesEnum.available.toString()) || rs.getType().getType().equals(AvailabilityTypesEnum.maybe.toString()))
+                                && rs.getType().getType().equals(AvailabilityTypesEnum.available) || rs.getType().getType().equals(AvailabilityTypesEnum.maybe))
                         .collect(Collectors.toCollection(ArrayList::new))
                         .size() > 0)
                 .collect(Collectors.toCollection(ArrayList::new));
