+++ src/test/java/resourceTests/PairResourceTests/PairResourceFindPairsInUnorderedSlots.java
@@ -1,4 +1,4 @@
-package resources;
+package resourceTests.PairResourceTests;
 
 import dao.SlotsDao;
 import domain.*;
@@ -7,9 +7,9 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import resources.PairResource;
 
 import java.sql.Date;
-import java.sql.Time;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.time.LocalTime;
@@ -47,10 +47,10 @@ public class PairResourceFindPairsInUnorderedSlots {
         endDate = dateFormat.format(nextDate);
 
         AvailabilityTypes availabilityType = MockDataUtil.createAvailableType((long) 1, AvailabilityTypesEnum.available);
-        SlotsTimes sameSlotsTimesFirst = MockDataUtil.createSlotTime((long) 2, LocalTime.of(8, 0, 0), LocalTime.of(9, 00, 0));
-        SlotsTimes sameSlotsTimesSecond = MockDataUtil.createSlotTime((long) 1, LocalTime.of(8, 30, 0), LocalTime.of(8, 30, 0));
-        SlotsTimes sameSlotsTimeThird = MockDataUtil.createSlotTime((long) 5, LocalTime.of(9, 0, 0), LocalTime.of(10, 30, 0));
-        SlotsTimes sameSlotsTimeFourth = MockDataUtil.createSlotTime((long) 3, LocalTime.of(9, 30, 0), LocalTime.of(10, 0, 0));
+        SlotsTimes sameSlotsTimesFirst = MockDataUtil.createSlotTime((long) 2, LocalTime.of(8, 30, 0), LocalTime.of(9, 0, 0));
+        SlotsTimes sameSlotsTimesSecond = MockDataUtil.createSlotTime((long) 1, LocalTime.of(8, 0, 0), LocalTime.of(8, 30, 0));
+        SlotsTimes sameSlotsTimeThird = MockDataUtil.createSlotTime((long) 5, LocalTime.of(10, 0, 0), LocalTime.of(10, 30, 0));
+        SlotsTimes sameSlotsTimeFourth = MockDataUtil.createSlotTime((long) 3, LocalTime.of(9, 0, 0), LocalTime.of(9, 30, 0));
         SlotsTimes sameSlotsTimeFifth = MockDataUtil.createSlotTime((long) 9, LocalTime.of(12, 0, 0), LocalTime.of(12, 30, 0));
         SlotsTimes sameSlotsTimeSixth = MockDataUtil.createSlotTime((long) 7, LocalTime.of(11, 0, 0), LocalTime.of(11, 30, 0));
         SlotsTimes sameSlotsTimeSeventh = MockDataUtil.createSlotTime((long) 8, LocalTime.of(11, 30, 0), LocalTime.of(12, 0, 0));
